PIPROD-1071: create new invoice component
