--- src/components/billPage/Invoice.css
@@ -1,258 +1,0 @@
-
-
-.navigation-invoice-login {
-  font-family: Rubik;
-  font-size: 28px;
-  font-weight: 700;
-  line-height: 37.92px;
-  color: #1a72dd;
-}
-
-.invoice-container {
-  padding: 20px;
-  max-width: 800px;
-}
-
-.invoice-header {
-  width: 100%;
-}
-
-.invoice-header-text {
-  margin-bottom: 20px;
-}
-
-.invoice-header-body {
-  width: 100%;
-  display: flex;
-  align-items: flex-start;
-  justify-content: space-between;
-}
-
-.invoice-header-left,
-.invoice-header-right {
-  width: 50%;
-  text-align: left;
-
-  .invoice-header-left-p,
-  .invoice-header-right-p {
-    font-size: 14px;
-  }
-}
-
-.invoice-header,
-.invoice-body {
-  text-align: left;
-  margin-bottom: 20px;
-}
-
-.invoice-header-heading {
-  text-align: center;
-}
-
-.invoice-header h1,
-.invoice-body h2 {
-  margin: 10px 0;
-}
-
-.invoice-body table {
-  width: 100%;
-  border-collapse: collapse;
-  margin-top: 20px;
-}
-
-.invoice-body table th,
-.invoice-body table td {
-  border: 1px solid #ddd;
-  padding: 8px;
-  text-align: left;
-}
-
-.invoice-body table th {
-  background-color: #f2f2f2;
-}
-
-.invoice-body {
-  .invoice-body-bottom {
-    margin-top: 10px;
-    display: flex;
-    align-items: start;
-    justify-content: space-between;
-
-    .invoice-body-bottom-left {
-      width: 50%;
-
-      .invoice-body-bottom-left-p {
-        font-size: 16px;
-      }
-    }
-
-    .invoice-body-bottom-right {
-      width: 50%;
-
-    }
-  }
-}
-
-.print-btn-outer {
-  z-index: 1;
-  width: 100%;
-  position: fixed;
-  bottom: 20px;
-  display: flex;
-  gap: 15px;
-  align-items: center;
-  justify-content: center;
-}
-
-.print-button {
-  padding: 20px 60px;
-  background-color: #1a72dd;
-  color: white;
-  border-radius: 13px;
-  border: 0.2px solid gray;
-  font-size: 16px;
-}
-
-@media (max-width: 600px) {
-  .invoice-container {
-    padding: 10px;
-  }
-
-  .invoice-header,
-  .invoice-body {
-    text-align: left;
-  }
-
-  .invoice-header p,
-  .invoice-body p {
-    font-size: 14px;
-  }
-
-  .invoice-body table th,
-  .invoice-body table td {
-    font-size: 12px;
-    padding: 5px;
-  }
-}
-
-
-@media print {
-
-  .navigation-header,
-  .print-btn-outer {
-    display: none;
-  }
-
-  .invoice-container {
-    margin: 0;
-    padding: 0;
-    width: 100%;
-  }
-
-  .navigation-invoice-login,
-  .navigation-header-container {
-    display: none;
-    margin-top: 10vh;
-  }
-}
-
-.print-btn-outer {
-  text-align: center;
-  margin-top: 20px;
-}
-
-.print-button {
-  padding: 10px 20px;
-  font-size: 16px;
-  cursor: pointer;
-}
-
-.whatsapp-number-popup {
-  width: 100%;
-  height: 100%;
-  position: fixed;
-  z-index: 2;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  background: rgba(0, 0, 0, 0.16);
-  border-radius: 16px;
-  box-shadow: 0 4px 30px rgba(0, 0, 0, 0.1);
-  backdrop-filter: blur(2.2px);
-  -webkit-backdrop-filter: blur(2.2px);
-
-
-  .whatsapp-number-form-container {
-    position: relative;
-    width: 90vw;
-    padding: 60px 30px 30px;
-    border-radius: 16px;
-    background-color: #ffffff;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    flex-direction: column;
-    gap: 30px;
-
-    .close-popup-button {
-
-      position: absolute;
-      right: 15px;
-      top: 15px;
-      background-color: transparent;
-      color: #000;
-
-    }
-
-    .whatsapp-number-input-container {
-      height: 55px;
-      width: 100%;
-      display: flex;
-      align-items: center;
-      justify-content: space-between;
-      border-radius: 16px;
-      overflow: hidden;
-      background-color: #0000000D;
-
-
-      padding: 5px 15px;
-
-      select {
-        background-color: transparent;
-        border: none;
-        outline: none;
-        height: 100%;
-        width: 20%;
-        flex-shrink: 0;
-      }
-
-      input {
-        background-color: transparent;
-        height: 100%;
-        width: inherit;
-        border: none;
-        outline: none;
-        padding: 0 15px;
-      }
-    }
-
-    .submit-number-button-disabled,
-    .submit-number-button-enable {
-      width: 40vw;
-      height: 45px;
-      border: none;
-      border-radius: 12px;
-      color: #ffffff;
-    }
-
-    .submit-number-button-disabled {
-      background-color: #cccccc;
-
-    }
-
-    .submit-number-button-enable {
-      background-color: #1a72dd;
-
-    }
-  }
-}
